Version 0.15
Changed SRChooseGrp to default to the top-level syntax group instead of the translated group.  Using SRChooseGrp! will choose the translated group (which will effect more values).
Added a help file.
Removed SRHiGrpNot and replaced it with SRHiGrp!.
Updated the examples.
Most of this work was done by Dr. Chip (thanks!).