Version 7.0
Changed SRSeach.

It will first look for the next item that has the syntax group chosen via SRChooseHiGrp.

It will also take an optional regular expression and not only find that syntax group, but also match the regular expression.

Added SRSearch!, which will find the next item that is NOT what was chosen via SRChooseHiGrp.

Same applied with the regular expression, so if I put my cursor on a comment and
    :SRChooseHiGrp.

Then ran
    :SRSearch! something

It will find the word "something" that is NOT in a comment.